* 'engine-api' of https://github.com/arangodb/arangodb: (46 commits)
  Merged RocksDBPathBasedIndex into RocksDBVPackIndex.
  fix handling of externals
  fix toArray()
  fix returning documents
  Added hash-, skip list, persistent indexes
  harmonize filtering - filter on full filename.
  mini modifications
  standardize starting of servers for replication tests
  Edge index compliant with design doc
  another fix for #2404
  Moved „removeLargeRange“ method to commons
  Added length byte to IndexValue keys.
  fixed issue #2404
  Added drop to primary index
  Added drop() to indexes
  Fixes a coordinator crash in Clustered Traversal. In case of failover during query planning.
  Fixed edge index
  Expand on CORS documentation
  Let ArangoDB handle the allow-credentials header
  make the authentication tests run on directories
  ...

+++ arangod/Pregel/Algos/EffectiveCloseness/EffectiveCloseness.cpp
@@ -66,7 +66,7 @@ struct ECComputation : public VertexComputation<ECValue, int8_t, HLLCounter> {
     // because the number of reachable vertices stays the same
     // when the compute method is not invoked
     if (value->shortestPaths.size() < globalSuperstep()) {
-      uint32_t i = value->shortestPaths.size();
+      size_t i = value->shortestPaths.size();
       int32_t numReachable = value->shortestPaths.back();
       for (; i < globalSuperstep(); i++) {
         value->shortestPaths.push_back(numReachable);
